# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "女性向 AI 互動應用後端服務，提供智能對話、互動小說、情感陪伴等功能",
        "title": "Thewavess AI Core API",
        "termsOfService": "https://thewavess.ai/terms",
        "contact": {
            "name": "Thewavess AI Core Team",
            "url": "https://thewavess.ai",
            "email": "api@thewavess.ai"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://thewavess.ai/license"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/monitor/stats": {
            "get": {
                "description": "獲取系統詳細運行狀態，包括硬體信息、記憶體使用、資料庫狀態、外部服務配置等完整監控信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitor"
                ],
                "summary": "獲取系統詳細狀態",
                "responses": {
                    "200": {
                        "description": "獲取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.SystemStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "獲取系統狀態失敗",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "$ref": "#/definitions/models.APIError"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.DatabaseInfo": {
            "type": "object",
            "properties": {
                "connected": {
                    "type": "boolean"
                },
                "ping_latency": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RedisInfo": {
            "type": "object",
            "properties": {
                "connected": {
                    "type": "boolean"
                },
                "ping_latency": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RuntimeInfo": {
            "type": "object",
            "properties": {
                "gc_count": {
                    "type": "integer"
                },
                "goroutines": {
                    "type": "integer"
                },
                "last_gc": {
                    "type": "string"
                },
                "memory_usage": {
                    "type": "string"
                },
                "next_gc": {
                    "type": "string"
                }
            }
        },
        "handlers.ServicesStatus": {
            "type": "object",
            "properties": {
                "grok": {
                    "type": "string"
                },
                "openai": {
                    "type": "string"
                },
                "tts": {
                    "type": "string"
                }
            }
        },
        "handlers.SystemInfo": {
            "type": "object",
            "properties": {
                "architecture": {
                    "type": "string"
                },
                "go_version": {
                    "type": "string"
                },
                "num_cpu": {
                    "type": "integer"
                },
                "os": {
                    "type": "string"
                }
            }
        },
        "handlers.SystemStatsResponse": {
            "type": "object",
            "properties": {
                "database": {
                    "$ref": "#/definitions/handlers.DatabaseInfo"
                },
                "redis": {
                    "$ref": "#/definitions/handlers.RedisInfo"
                },
                "runtime": {
                    "$ref": "#/definitions/handlers.RuntimeInfo"
                },
                "services": {
                    "$ref": "#/definitions/handlers.ServicesStatus"
                },
                "status": {
                    "type": "string"
                },
                "system": {
                    "$ref": "#/definitions/handlers.SystemInfo"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_REQUEST"
                },
                "details": {
                    "type": "string",
                    "example": "用戶名稱不能為空"
                },
                "message": {
                    "type": "string",
                    "example": "請求參數無效"
                }
            }
        },
        "models.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/models.APIError"
                },
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "請輸入 'Bearer ' + JWT token，例如: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}